Merge was done with,
sdb-common-android-jb-4.2.2 reseted to tag MP_3.30.1_BCM21664Android_SystemRel_6.2.11.1

And,
aosp branch experimental/android-3.10 was at,

commit a88f9e27498afaea615ad3e93af4f26df1f84987
Author: Arve Hjønnevåg <arve@android.com>
Date:   Thu Jun 6 20:05:40 2013 -0700

    kconfig: Fix defconfig when one choice menu selects options that another choice menu depends on

    The defconfig and Kconfig combination below, which is based on 3.10-rc4
    Kconfigs, resulted in several options getting set to "m" instead of "y".

Conflicts:
	Documentation/devicetree/bindings/arm/arch_timer.txt
	arch/arm/configs/bcm21664_hawaii_garnet_edn010_defconfig
	arch/arm/configs/bcm21664_hawaii_garnet_edn020_defconfig
	arch/arm/configs/bcm23550_java_edn000_defconfig
	arch/arm/include/asm/arch_timer.h
	arch/arm/include/asm/dma-mapping.h
	arch/arm/kernel/arch_timer.c
	arch/arm/kernel/smp.c
	arch/arm/lib/delay.c
	arch/arm/mach-java/board-garnet.c
	arch/arm/mm/dma-mapping.c
	arch/arm/mm/mm.h
	arch/arm/mm/mmap.c
	arch/arm/plat-kona/include/mach/uncompress.h
	arch/arm/plat-kona/kona_pm.c
	drivers/Kconfig
	drivers/char/broadcom/mm/common/mm_common.c
	drivers/char/broadcom/mm/common/mm_dvfs.h
	drivers/clocksource/Kconfig
	drivers/clocksource/Makefile
	drivers/clocksource/arm_arch_timer.c
	drivers/cpufreq/cpufreq.c
	drivers/cpufreq/cpufreq_conservative.c
	drivers/cpufreq/cpufreq_ondemand.c
	drivers/gpu/ion/ion.c
	drivers/gpu/ion/ion_page_pool.c
	drivers/media/platform/soc_camera/soc_camera.c
	drivers/media/video/Kconfig
	drivers/media/video/Makefile
	drivers/misc/pmem.c
	drivers/mmc/card/block.c
	drivers/staging/android/Kconfig
	drivers/staging/android/lowmemorykiller.c
	drivers/usb/gadget/f_mass_storage.c
	drivers/usb/otg/bcmpmu59xxx_otg_xceiv.c
	drivers/video/backlight/pwm_bl.c
	include/clocksource/arm_arch_timer.h
	include/linux/dma-contiguous.h
	include/linux/mm.h
	include/linux/mm_inline.h
	include/linux/mmzone.h
	mm/oom_kill.c
	mm/page_alloc.c
	mm/vmscan.c

+++ drivers/char/broadcom/modem/rpc/rpc_client_msgs.h
@@ -18,7 +18,7 @@
 
 int rpc_is_registered_msg(UInt16 dscm, UInt8 clientId);
 Boolean rpc_register_client_msgs(UInt8 clientId, UInt16 *tbl,
-					UInt16 table_size);
+					UInt32 table_size);
 Boolean rpc_reset_client_msgs(UInt8 clientId);
